Refactoring for an extensible Index API: Part 3
- Move CoveringIndex specific code into the
  com.microsoft.hyperspace.index.types.covering package
- Move traits/classes/objects in the
  com.microsoft.hyperspace.index.rules package into their own files
- Add alias for CoveringIndexConfig for backward compatibility
