+++ src/test/scala/com/microsoft/hyperspace/index/IndexCollectionManagerTest.scala
@@ -25,6 +25,7 @@ import org.mockito.Mockito.{mock, when}
 import com.microsoft.hyperspace.HyperspaceException
 import com.microsoft.hyperspace.actions.Constants
 import com.microsoft.hyperspace.index.IndexConstants.{REFRESH_MODE_FULL, REFRESH_MODE_INCREMENTAL}
+import com.microsoft.hyperspace.index.types.covering.CoveringIndex
 
 class IndexCollectionManagerTest extends HyperspaceSuite {
   private val testLogManagerFactory: IndexLogManagerFactory = new IndexLogManagerFactory {
