+++ src/test/scala/com/microsoft/hyperspace/index/IndexTest.scala
@@ -20,6 +20,7 @@ import org.apache.spark.SparkFunSuite
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
 
 import com.microsoft.hyperspace.actions.Constants
+import com.microsoft.hyperspace.index.types.covering.CoveringIndex
 
 class IndexTest extends SparkFunSuite {
   val indexConfig1 = IndexConfig("myIndex1", Array("id"), Seq("name"))
