+++ src/test/scala/com/microsoft/hyperspace/util/JsonUtilsTest.scala
@@ -21,6 +21,7 @@ import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructT
 
 import com.microsoft.hyperspace.actions.Constants
 import com.microsoft.hyperspace.index._
+import com.microsoft.hyperspace.index.types.covering.CoveringIndex
 
 class JsonUtilsTest extends SparkFunSuite {
   test("Test for JsonUtils.") {
